Moves get_abi_type_str to it own file.
