+++ sway-core/src/abi_generation/mod.rs
@@ -1,2 +1,3 @@
+pub mod abi_str;
 pub mod evm_abi;
 pub mod fuel_abi;
